ctci-making-anagrams/src/main/java/Solution.java
/*
 * Strings: Making Anagrams
 *
 * https://www.hackerrank.com/challenges/ctci-making-anagrams/problem?h_l=interview&playlist_slugs%5B%5D=interview-preparation-kit&playlist_slugs%5B%5D=strings
 */

import java.io.BufferedWriter;
import java.io.FileWriter;
import java.io.IOException;
import java.util.Arrays;
import java.util.Scanner;

public class Solution {

    // Complete the makeAnagram function below.
    static int makeAnagram(String a, String b) {
        final int[] charsFrequency = new int[26];

        a.chars().forEach(c -> {
            charsFrequency[c - 97]++;
        });

        b.chars().forEach(c -> {
            charsFrequency[c - 97]--;
        });

       return Arrays.stream(charsFrequency).reduce(0, (x, y) -> x + Math.abs(y));
    }

    private static final Scanner scanner = new Scanner(System.in);

    public static void main(String[] args) throws IOException {
        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(System.getenv("OUTPUT_PATH")));

        String a = scanner.nextLine();

        String b = scanner.nextLine();

        int res = makeAnagram(a, b);

        bufferedWriter.write(String.valueOf(res));
        bufferedWriter.newLine();

        bufferedWriter.close();

        scanner.close();
    }
}
